requirements.txt
backcall==0.1.0
cycler==0.10.0
decorator==4.4.1
Django==2.2.7
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
opencv-python==4.1.2.30
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.5
pytesseract==0.3.0
python-dateutil==2.8.1
pytz==2019.3
scipy==1.3.2
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
